improved way self-referencing lines are handled by the level loader.
This deals with the way we determine which sector a subsector should
belong to.  When the only real seg(s) in a GL-NODE subsector are on
self-ref linedefs (which have same sector on both sides), the new logic
looks to grab the sector from a neighbor subsector (which in turn may
grab it from another neighbor).