DEH_EDGE : ignore MT_SPAWNFIRE when creating attacks.
The old code did this (in a very hacky way) and I accidentally broke it
with the recent rework of the DEH_EDGE code.

This commit "fixes" the fatal errors given by REKKR (etc), but I think
more work is needed to get DEHACKED mods using the brain-spit stuff to
behave exactly the same as in Vanilla and Boom/MBF ports.