removed the macro redefinitions for `strcmp`, `strncmp`, `stricmp`.
I thought this was a bit dodgy, and yes indeed it did come to bite me on
the arse, since some parts of the SDL2 headers in the MSYS build got upset
and didn't want to play ball.

Hence these versions of `strcmp` and `strncmp` are in the epi namespace,
and I have replaced `stricmp` (which is non-standard) throughout the code
with the `epi::case_cmp` function.